# Emit classes