--- a/baselines/webworker.generated.d.ts
+++ b/baselines/webworker.generated.d.ts
@@ -152,7 +152,7 @@ declare class Blob {
     msClose(): void;
     msDetachStream(): any;
     slice(start?: number, end?: number, contentType?: string): Blob;
-    new (blobParts?: any[], options?: BlobPropertyBag): Blob;
+    constructor(blobParts?: any[], options?: BlobPropertyBag);
 }
 
 declare class Cache {
@@ -261,7 +261,7 @@ declare class DOMException {
     readonly URL_MISMATCH_ERR: number;
     readonly VALIDATION_ERR: number;
     readonly WRONG_DOCUMENT_ERR: number;
-    new(message?: string, name?: string): DOMException;
+    constructor(message?: string, name?: string);
     static readonly ABORT_ERR: number;
     static readonly DATA_CLONE_ERR: number;
     static readonly DOMSTRING_SIZE_ERR: number;
@@ -306,7 +306,7 @@ declare class ErrorEvent {
     readonly lineno: number;
     readonly message: string;
     initErrorEvent(typeArg: string, canBubbleArg: boolean, cancelableArg: boolean, messageArg: string, filenameArg: string, linenoArg: number): void;
-    new(type: string, errorEventInitDict?: ErrorEventInit): ErrorEvent;
+    constructor(type: string, errorEventInitDict?: ErrorEventInit);
 }
 
 declare class Event {
@@ -349,7 +349,7 @@ declare class File {
     readonly name: string;
     readonly webkitRelativePath: string;
     readonly lastModified: number;
-    new (parts: (ArrayBuffer | ArrayBufferView | Blob | string)[], filename: string, properties?: FilePropertyBag): File;
+    constructor(parts: (ArrayBuffer | ArrayBufferView | Blob | string)[], filename: string, properties?: FilePropertyBag);
 }
 
 declare class FileList {
@@ -384,7 +384,7 @@ declare class Headers {
     get(name: string): string | null;
     has(name: string): boolean;
     set(name: string, value: string): void;
-    new(init?: HeadersInit): Headers;
+    constructor(init?: HeadersInit);
 }
 
 declare class IDBCursor {
@@ -557,8 +557,8 @@ declare class ImageData {
     data: Uint8ClampedArray;
     readonly height: number;
     readonly width: number;
-    new(width: number, height: number): ImageData;
-    new(array: Uint8ClampedArray, width: number, height: number): ImageData;
+    constructor(width: number, height: number);
+    constructor(array: Uint8ClampedArray, width: number, height: number);
 }
 
 declare class MessageChannel {
@@ -574,7 +574,7 @@ declare class MessageEvent {
     readonly ports: any;
     readonly source: any;
     initMessageEvent(typeArg: string, canBubbleArg: boolean, cancelableArg: boolean, dataArg: any, originArg: string, lastEventIdArg: string, sourceArg: any): void;
-    new(type: string, eventInitDict?: MessageEventInit): MessageEvent;
+    constructor(type: string, eventInitDict?: MessageEventInit);
 }
 
 interface MessagePortEventMap {
@@ -703,7 +703,7 @@ declare class ProgressEvent {
     readonly loaded: number;
     readonly total: number;
     initProgressEvent(typeArg: string, canBubbleArg: boolean, cancelableArg: boolean, lengthComputableArg: boolean, loadedArg: number, totalArg: number): void;
-    new(type: string, eventInitDict?: ProgressEventInit): ProgressEvent;
+    constructor(type: string, eventInitDict?: ProgressEventInit);
 }
 
 declare class PushManager {
@@ -768,7 +768,7 @@ declare class Response {
     readonly url: string;
     readonly redirected: boolean;
     clone(): Response;
-    new(body?: any, init?: ResponseInit): Response;
+    constructor(body?: any, init?: ResponseInit);
     error: () => Response;
     redirect: (url: string, status?: number) => Response;
 }
@@ -831,7 +831,7 @@ declare class URL {
     username: string;
     readonly searchParams: URLSearchParams;
     toString(): string;
-    new(url: string, base?: string | URL): URL;
+    constructor(url: string, base?: string | URL);
     static createObjectURL(object: any, options?: ObjectURLOptions): string;
     static revokeObjectURL(url: string): void;
 }