This PR converts interface + variables to interface + classes.

Original:

```ts
interface AudioProcessingEvent extends Event {
    readonly inputBuffer: AudioBuffer;
    readonly outputBuffer: AudioBuffer;
    readonly playbackTime: number;
}

declare var AudioProcessingEvent: {
    prototype: AudioProcessingEvent;
    new(): AudioProcessingEvent;
}
```

This PR:

```ts
interface AudioProcessingEvent extends Event {}
declare class AudioProcessingEvent {
    readonly inputBuffer: AudioBuffer;
    readonly outputBuffer: AudioBuffer;
    readonly playbackTime: number;
}
```

---

Problem:

This looks performance-wise worse than original variable emits.

Original:

```
Files:            2
Lines:        19316
Nodes:        84450
Identifiers:  29164
Symbols:      22004
Types:         5790
Memory used: 54963K
I/O read:     0.00s
I/O write:    0.00s
Parse time:   0.38s
Bind time:    0.15s
Check time:   0.60s
Emit time:    0.02s
Total time:   1.15s
```

This PR:

```
Files:            2
Lines:        17121
Nodes:        79099
Identifiers:  26931
Symbols:      78473
Types:        21731
Memory used: 90920K
I/O read:     0.00s
I/O write:    0.00s
Parse time:   0.35s
Bind time:    0.17s
Check time:   1.04s
Emit time:    0.02s
Total time:   1.58s
```

I haven't digged more but I suspect declaration merging is bad. (This is done because of Microsoft/TypeScript#5468)